fix 32 bit numbers

+++ include/ast.hh
@@ -74,9 +74,9 @@ struct NodeAssign : public Node {
     Node for integer literals
 */
 struct NodeInt : public Node {
-    long value;
+    long long value;
 
-    NodeInt(long val);
+    NodeInt(long long val);
     std::string to_string();
     llvm::Value *llvm_codegen(LLVMCompiler *compiler);
 };
